create base class
creation of base class and update color_light to inherit that
